CCS-855 moves CW errors to the shared library
