+++ src/errors/index.js
@@ -1,5 +1,8 @@
 const idvErrors = require('./idvErrors');
+const { ErrorCodes, ErrorContextTypes } = require('./definitions');
 
 module.exports = {
-  idvErrors,
+  ErrorCodes,
+  ErrorContextTypes,
+  idvErrors, // For retrofit, will be deprecated on the future
 };
